# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CHPM Survey API
  version: 1.0.0
  summary: Role- and scope-aware questionnaire collection API
  description: |
    REST API for internal staff authentication, scoped administration, questionnaire
    authoring/versioning, invitation delivery, token-based respondent sessions, managed
    terminals, pseudonymized statistics, compliance operations, notifications, audit,
    and the controlled exceptional legal-access workflow.

    The specification describes implemented NestJS controller operations only. Direct
    respondent contact data is not available through normal business endpoints. The
    exceptional DPO export is allowed only after independent DPO and legal validation,
    is restricted to explicit public codes, returns an AES-256-GCM encrypted envelope,
    expires automatically, and is fully audited. The local DPO console enforces the same
    approved-request prerequisite.

    Version 1.0 is the production handover contract. Compatibility changes follow the
    release and upgrade procedure in `docs/production/installation.md`.
  contact:
    name: CHPM Survey maintainers
  license:
    name: Proprietary / internal project
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
    description: Same-origin reverse-proxy deployment
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: chpm_session
      description: Opaque internal staff session. Only a SHA-256 hash is stored server-side.
    respondentQueryToken:
      type: apiKey
      in: query
      name: token
      description: Signed secret bound to one respondent invitation/session.
    terminalQueryToken:
      type: apiKey
      in: query
      name: token
      description: Secret managed-device token used by the public terminal session endpoint.
    metricsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Dedicated monitoring token supplied through `METRICS_TOKEN`; not a staff session.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      description: Accepted or generated request/correlation identifier.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    BadRequest:
      description: Invalid payload, query, path value, or workflow request.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        X-Correlation-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, revoked, or inactive session/token.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated caller lacks the required role or object scope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Object does not exist or is intentionally hidden because it is outside the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Duplicate or invalid lifecycle transition, including immutable/already-submitted state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [statusCode, error, path, timestamp]
      properties:
        statusCode:
          type: integer
          minimum: 400
          maximum: 599
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              examples: [VALIDATION_ERROR]
            message:
              type: string
            details:
              type: array
              items:
                type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
...
    UserRole:
      type: string
      enum:
        - admin
        - site_manager
        - moderator
        - questionnaire_admin
        - analyst
        - dpo
        - judicial_officer
        - technical_admin
        - service_account
        - respondent
...
    Building:
      type: object
      required: [id, code, label, city, country, timezone]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        label:
          type: string
        city:
          type: string
        country:
          type: string
        timezone:
          type: string
          examples: [America/Toronto]
        siteId:
          type: [string, 'null']
          format: uuid
        organizationId:
          type: [string, 'null']
          format: uuid
    BuildingsResponse:
      type: object
      additionalProperties: false
      required: [buildings]
      properties:
        buildings:
          type: array
          items:
            $ref: '#/components/schemas/Building'
    CreateBuildingRequest:
      type: object
      additionalProperties: false
      required: [code, label, city, country, timezone]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 40
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
        label:
          type: string
          minLength: 2
          maxLength: 160
        city:
          type: string
          minLength: 2
          maxLength: 100
        country:
          type: string
          minLength: 2
          maxLength: 100
        timezone:
          type: string
          minLength: 3
          maxLength: 100
    BuildingResponse:
      type: object
      additionalProperties: false
      required: [building]
      properties:
        building:
          $ref: '#/components/schemas/Building'
    OrganizationSummary:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    Site:
      type: object
      required: [id, code, name, organizationId]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        organization:
          oneOf:
            - $ref: '#/components/schemas/OrganizationSummary'
            - type: 'null'
        country:
          type: [string, 'null']
        timezone:
          type: [string, 'null']
    SitesResponse:
      type: object
      additionalProperties: false
      required: [sites]
      properties:
        sites:
          type: array
          items:
            $ref: '#/components/schemas/Site'
    CreateSiteRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 40
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
        name:
          type: string
          minLength: 2
          maxLength: 160
        country:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 100
    SiteResponse:
      type: object
      additionalProperties: false
      required: [site]
      properties:
        site:
          $ref: '#/components/schemas/Site'
    SiteTeamUser:
      type: object
      required: [id, email, displayName, role, isActive, organizationId, siteId, buildingId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Internal staff account email.
        displayName:
          type: string
        role:
          type: string
          enum: [site_manager, moderator]
        roleLabel:
          type: string
        isActive:
          type: boolean
        organizationId:
          type: [string, 'null']
          format: uuid
        siteId:
          type: [string, 'null']
          format: uuid
        buildingId:
          type: [string, 'null']
          format: uuid
        site:
          oneOf:
            - $ref: '#/components/schemas/Site'
            - type: 'null'
        building:
          oneOf:
            - $ref: '#/components/schemas/Building'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SiteAdminsResponse:
      type: object
      required: [users, policy]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/SiteTeamUser'
        policy:
          type: object
          required: [manageableRoles, scope, passwordReturnedOnce, forbiddenRoles]
          properties:
            manageableRoles:
              type: array
              items:
                const: site_manager
            scope:
              const: project
            passwordReturnedOnce:
              const: true
            forbiddenRoles:
              type: array
              items:
                $ref: '#/components/schemas/UserRole'
    SiteTeamResponse:
      type: object
      required: [users, policy]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/SiteTeamUser'
        policy:
          type: object
          required: [manageableRoles, scope, passwordReturnedOnce]
          properties:
            manageableRoles:
              type: array
              items:
                const: moderator
            scope:
              const: site
            passwordReturnedOnce:
              const: true
...
    ProjectHierarchyResponse:
      type: object
      additionalProperties: false
      required: [hierarchy, scope, generatedAt]
      properties:
        hierarchy:
          $ref: '#/components/schemas/ProjectHierarchyNode'
        scope:
          type: string
          enum: [project, site, self]
        generatedAt:
          type: string
          format: date-time
    CreateSiteAdminRequest:
      type: object
      additionalProperties: false
      required: [email, displayName, siteId]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        siteId:
          type: string
          format: uuid
        temporaryPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
    UpdateSiteAdminRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        siteId:
          type: string
          format: uuid
        isActive:
          type: boolean
    CreateSiteModeratorRequest:
      type: object
      additionalProperties: false
      required: [email, displayName, buildingId]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        buildingId:
          type: string
          format: uuid
        temporaryPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
    UpdateSiteModeratorRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        buildingId:
          type: string
          format: uuid
        isActive:
          type: boolean
    UserMutationResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/SiteTeamUser'
        temporaryPassword:
          type: string
          format: password
          readOnly: true
          description: Returned once when generated/reset. Never log or persist in application storage.
        temporaryPasswordGenerated:
          type: boolean
    RevokeSessionsResponse:
      type: object
      additionalProperties: false
      required: [user, revokedSessionCount]
      properties:
        user:
          $ref: '#/components/schemas/SiteTeamUser'
        revokedSessionCount:
          type: integer
          minimum: 0
...
paths:
...
  /buildings:
    get:
      tags: [Buildings]
      operationId: listScopedBuildings
      summary: List buildings visible to the current role and scope
      x-required-roles: [admin, moderator, site_manager, questionnaire_admin, analyst]
      responses:
        '200':
          description: Scoped building list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/sites:
    get:
      tags: [Project administration]
      operationId: listProjectSites
      summary: List sites managed by the project administrator
      description: Returns project reference data only and no respondent identity data.
      x-required-roles: [admin]
      responses:
        '200':
          description: Sites in the administrator's organization/project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SitesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Project administration]
      operationId: createProjectSite
      summary: Create a site in the administrator's organization
      x-required-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '201':
          description: Site created in the caller's organization and audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/site-admins:
    get:
      tags: [Project administration]
      operationId: listSiteManagers
      summary: List site-manager accounts
      x-required-roles: [admin]
      responses:
        '200':
          description: Scoped site-manager accounts and delegation policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAdminsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Project administration]
      operationId: createSiteManager
      summary: Create or provision a site-manager account
      description: Cannot create project administrators, DPO, judicial, or technical roles.
      x-required-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteAdminRequest'
      responses:
        '201':
          description: Site manager created; a generated temporary password may be returned once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/site-admins/{id}:
    patch:
      tags: [Project administration]
      operationId: updateSiteManager
      summary: Update a site manager's name, site, or active state
      x-required-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteAdminRequest'
      responses:
        '200':
          description: Site manager updated; sensitive changes revoke sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/site-admins/{id}/reset-password:
    post:
      tags: [Project administration]
      operationId: resetSiteManagerPassword
      summary: Reset a site manager's password and revoke sessions
      x-required-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Temporary password returned once and prior sessions revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/site-admins/{id}/revoke-sessions:
    post:
      tags: [Project administration]
      operationId: revokeSiteManagerSessions
      summary: Revoke every active session for a site manager
      x-required-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Sessions revoked and action audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeSessionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /site/buildings:
    post:
      tags: [Site administration]
      operationId: createSiteBuilding
      summary: Create a building in the site manager's assigned site
      x-required-roles: [site_manager]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '201':
          description: Building created in the assigned site and organization and audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /site/team:
    get:
      tags: [Site administration]
      operationId: listSiteTeam
      summary: List moderator accounts for the current site
      x-required-roles: [site_manager]
      responses:
        '200':
          description: Moderators in the current site and delegation policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteTeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /site/moderators:
    post:
      tags: [Site administration]
      operationId: createOrReactivateSiteModerator
      summary: Create or reactivate a moderator in the current site
      x-required-roles: [site_manager]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteModeratorRequest'
      responses:
        '201':
          description: Moderator created/reactivated; temporary password may be returned once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /site/moderators/{id}:
    patch:
      tags: [Site administration]
      operationId: updateSiteModerator
      summary: Update a moderator in the current site
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteModeratorRequest'
      responses:
        '200':
          description: Moderator updated; sensitive changes revoke sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /site/moderators/{id}/reset-password:
    post:
      tags: [Site administration]
      operationId: resetSiteModeratorPassword
      summary: Reset a moderator password and revoke sessions
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Temporary password returned once and prior sessions revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /site/moderators/{id}/revoke-sessions:
    post:
      tags: [Site administration]
      operationId: revokeSiteModeratorSessions
      summary: Revoke every active session for a moderator
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Sessions revoked and action audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeSessionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/hierarchy:
    get:
      tags: [Site administration]
      operationId: getProjectHierarchy
      summary: Return the caller-visible operational project hierarchy
      description: |
        Builds a collapsible project tree without exposing staff emails. Project administrators
        see every operational user in their organization; site managers see project administrators,
        their own site, themselves, and their moderators; moderators see only their upward chain and themselves.
      x-required-roles: [admin, site_manager, moderator]
      responses:
        '200':
          description: Role- and scope-filtered project hierarchy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectHierarchyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/site-team:
    get:
      tags: [Site administration]
      operationId: listSiteTeamLegacy
      summary: List the current site team through the legacy alias
      deprecated: true
      description: Use `GET /site/team` for new clients.
      x-required-roles: [site_manager]
      responses:
        '200':
          description: Moderators in the current site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteTeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/site-moderators:
    post:
      tags: [Site administration]
      operationId: createOrReactivateSiteModeratorLegacy
      summary: Create or reactivate a moderator through the legacy alias
      deprecated: true
      description: Use `POST /site/moderators` for new clients.
      x-required-roles: [site_manager]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteModeratorRequest'
      responses:
        '201':
          description: Moderator created/reactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/site-moderators/{id}:
    patch:
      tags: [Site administration]
      operationId: updateSiteModeratorLegacy
      summary: Update a moderator through the legacy alias
      deprecated: true
      description: Use `PATCH /site/moderators/{id}` for new clients.
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteModeratorRequest'
      responses:
        '200':
          description: Moderator updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/site-moderators/{id}/reset-password:
    post:
      tags: [Site administration]
      operationId: resetSiteModeratorPasswordLegacy
      summary: Reset a moderator password through the legacy alias
      deprecated: true
      description: Use `POST /site/moderators/{id}/reset-password` for new clients.
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Temporary password returned once and sessions revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/site-moderators/{id}/revoke-sessions:
    post:
      tags: [Site administration]
      operationId: revokeSiteModeratorSessionsLegacy
      summary: Revoke moderator sessions through the legacy alias
      deprecated: true
      description: Use `POST /site/moderators/{id}/revoke-sessions` for new clients.
      x-required-roles: [site_manager]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Sessions revoked and action audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeSessionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...